psusensor: fixed not activate for multiple power state
Link:https://gerrit.openbmc.org/c/openbmc/dbus-sensors/+/77426

Due to the deactivation function being sensor-specific
while the check findI2CDev != devices.end() is chip-specific,
some sensors may fail to activate during power state changes.

For example, consider the following EM configuration:

{
  "in0_Name": "MB_P12V_MEM0_VOLT_V",
  "in0_PowerState": "Always",
  "in1_Name": "MB_P5V_STBY_VOLT_V",
  "in2_Name": "MB_PVDD33_S5_VOLT_V",
  "in3_Name": "MB_P1V8_STBY_VOLT_V",
  "in4_Name": "MB_P12V_MEM1_VOLT_V",
  "in5_Name": "MB_P1V8_CPU_RT_VOLT_V",
  "in5_PowerState": "On"
}

If in0_PowerState is set to Always, and we perform a DC off followed
by a DC on, in5 might fail to activate. This happens because
the condition "activateOnly && !findI2CDev->second.second"
evaluates to true, causing the function to return early.

To address this issue, an additional condition has been added to bypass
the check when multiple power states are detected, ensuring that all
sensors can be properly activated

Change-Id: I6d91a8f2d5c86e215bd89f8d03bee6f8ecdc502a
Signed-off-by: Peter Yin <peter.yin@quantatw.com>